requirements.txt
Flask==2.3.3
PyYAML==6.0.2
typing-extensions==4.14.1
pytest>=7.0.0
pytest-asyncio==1.1.0
pydantic-settings>=2.0.0
pydantic>=2,<3
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
PyJWT>=2.8.0

numpy==1.26.4

hypothesis==6.138.13
python-dotenv==1.1.1
# Optional FS hot-reload:
# watchdog>=3.0.0
